moved SymbolicData instances from ArithmeticCircuit.Instance
